Refactor general
